lua_tools moved to xrLua. R2 compilation temporary fix.
