+++ xrShared/xrShared.vcxproj
@@ -156,11 +156,9 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="lua_tools.cpp" />
     <ClCompile Include="stdafx.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="lua_tools.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="xrShared.h" />
   </ItemGroup>
